Clone form options before setting to res.locals
This means that the locals options can be mutated in request pipelines without also mutating the options for all requests.

The intention for this is that the template mixins can use `locals.options.fields` to get field configuration instead of being passed a fixed fields object on instantiation, and implementing controllers can modify field config on a per-request basis.