Use current error object properties when rendering message
Rather than taking the properties from the arguments when the error was created, allow the `getMessage` method to use the current properties of an error instance when getting its message property.

This allows controllers to modify the properties of an error after its instantiation in order to implement custom error behaviour.

In particular, in one app we have two different fields which write to the same data point on the session, but from different places depending on context (address looku vs manual address entry). Depending on the context, the error message for the `required` validator on that field needs to be different. If the controller can override the `key` property of that error without having to actually change the name of the field then it can display custom error messages.